requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
datasets==2.14.2
dill==0.3.7
Django==4.2.3
efficientnet-pytorch==0.7.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
huggingface-hub==0.16.4
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.10.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.25.1
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
